style: directory refactoring
