+++ src/main/java/sparta/a08/trello/common/security/UserDetailsService.java
@@ -1,9 +1,11 @@
-package sparta.a08.trello.user;
+package sparta.a08.trello.common.security;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import sparta.a08.trello.user.entity.User;
+import sparta.a08.trello.user.repository.UserRepository;
 
 @Service
 @RequiredArgsConstructor
