+++ src/main/java/sparta/a08/trello/user/dto/UserResponseDTO.java
@@ -1,6 +1,7 @@
-package sparta.a08.trello.user;
+package sparta.a08.trello.user.dto;
 
 import lombok.Getter;
+import sparta.a08.trello.user.entity.User;
 
 @Getter
 public class UserResponseDTO {
