+++ src/test/java/sparta/a08/trello/common/CommonTest.java
@@ -1,7 +1,7 @@
 package sparta.a08.trello.common;
 
 
-import sparta.a08.trello.user.User;
+import sparta.a08.trello.user.entity.User;
 
 public interface CommonTest {
     Long TEST_USER_ID = 1L;
